Task lists and Tablet Support
Can now add multiple lists of tasks. On a tablet (or other large-screen device), the list of task lists will be displayed on the left while the current set of tasks are on the right. On a phone, it will default to the list of task lists. Selecting a list will bring up a fragment with the list items.